+++ src/Calculator/Operations/Basic/DivideOperation.cs
@@ -7,7 +7,7 @@ public class DivideOperation : BasicOperation
     {
     }
 
-    public override decimal Calculate()
+    public override double Calculate()
     {
         ArgumentNullException.ThrowIfNull(LeftOperand);
         ArgumentNullException.ThrowIfNull(RightOperand);
